% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedcldlist.R
\name{orderedcldlist}
\alias{orderedcldlist}
\title{: Ordered cldList}
\usage{
orderedcldlist(
  data = NULL,
  comparison = NULL,
  p.value = NULL,
  threshold = 0.05,
  print.comp = FALSE,
  desired_order = NULL
)
}
\arguments{
\item{data}{Tukey HSD results, formatted as shown in the example below}

\item{comparison}{A vector of text describing comparisons, with each element in the form "Treat.A-Treat.B"}

\item{p.value}{A vector of p-values corresponding to the comparisons in the comparison argument}

\item{threshold}{The alpha value: the p-value below which the comparison will be considered statistically significant}

\item{print.comp}{If TRUE, prints out a data frame of the comparisons. Useful for debugging}

\item{desired_order}{A vector of one or more comparisons in the desired order, e.g. c("Treat.B", "Treat.A")}
}
\value{
A list of variables with a letter denoting statistical significance
}
\description{
: This function takes the output of TukeyHSD post-hoc test, modified to include
a Comparison column
}
\examples{
set.seed(12345)
df <- data.frame(program = rep(c("AA", "BB", "CC", "DD"), each = 10),
                 weight_loss = c(runif(10, 0, 10),
                                 runif(10, 0, 4),
                                 runif(10, 1, 7),
                                 runif(10, 0, 3)),
                 group = factor(sample(c(0,1), replace = TRUE, size = 10)))
model = aov(weight_loss ~ program * group, data = df)
summary(model)
TUK = TukeyHSD(model, ordered = TRUE)
TUK = as.data.frame(TUK[[1]])
HSD = data.frame(Comparison=row.names(TUK),
                 diff=TUK$diff, lwr=TUK$lwr, upr=TUK$upr, p.adj=TUK$`p adj`)
HSD
desired_order = c("BB", "AA", "CC", "DD")
orderedcldlist(data = HSD,
               comparison = "Comparison",
               p.value = "p.adj",
               threshold = 0.05,
               desired_order = desired_order)

desired_order2 = c("DD", "BB", "CC", "AA")
orderedcldlist(data = HSD,
               comparison = "Comparison",
               p.value = "p.adj",
               threshold = 0.05,
               desired_order = desired_order2)

}
